training/requirements.txt
pandas
google-cloud-storage
scikit-learn
